fix references to service utils
